+++ web/src/pages/profile/components/ProfileHeader.tsx
@@ -1,0 +1,10 @@
+export default function ProfileHeader() {
+  return (
+    <div>
+      <h1 className="text-2xl font-semibold tracking-tight">Profile</h1>
+      <p className="text-sm text-muted-foreground mt-1">
+        Manage your account and see your event participation history.
+      </p>
+    </div>
+  );
+}
