+++ web/src/pages/profile/types/Profile.ts
@@ -1,0 +1,12 @@
+export interface ParticipationEvent {
+  documentId: string;
+  title: string;
+  start?: string;
+  end?: string;
+}
+
+export interface ParticipationRecord {
+  documentId: string;
+  registeredAt?: string;
+  event?: ParticipationEvent;
+}
